# perf: Fix false results and add Fedora support