--- a/perf/perf_duplicate_probe.py
+++ b/perf/perf_duplicate_probe.py
@@ -34,7 +34,7 @@ class PerfDuplicateProbe(Test):
         if 'Ubuntu' in distro_name:
             deps.extend(['linux-tools-common', 'linux-tools-%s' %
                          platform.uname()[2]])
-        elif distro_name in ['rhel', 'SuSE']:
+        elif distro_name in ['rhel', 'SuSE', 'fedora', 'centos']:
             deps.extend(['perf'])
         else:
             self.cancel("Install the package for perf supported\