--- a/perf/perf_script_bug.py.data/probe.cfg
+++ b/perf/perf_script_bug.py.data/probe.cfg
@@ -4,3 +4,5 @@ probepoint = 3
 probepoint = 3
 [Ubuntu]
 probepoint = 3
+[fedora]
+probepoint = 3