### perf: Fix false results and add Fedora support

Updated perf tests to enable fedora testing

This patch will also enable "perf test" failures
which were earlier cancelled due to missing
"fedora" support in distro.detect()

Signed-off-by: Misbah Anjum N [misanjum@linux.vnet.ibm.com](mailto:misanjum@linux.vnet.ibm.com)