@misanjumn 2024-03-26
## Testing on OS: Fedora40
## Without the patch
CANCEL: 2
ERROR: 1
```
(1/3) /home/misanjumn/tests/tests/avocado-misc-tests/perf/perf_duplicate_probe.py:PerfDuplicateProbe.test_duplicate_probe: STARTED
 (1/3) /home/misanjumn/tests/tests/avocado-misc-tests/perf/perf_duplicate_probe.py:PerfDuplicateProbe.test_duplicate_probe:  CANCEL: Install the package for perf supported                      by fedora (0.77 s)
 (2/3) /home/misanjumn/tests/tests/avocado-misc-tests/perf/perf_script_bug.py:PerfScript.test_script_probe: STARTED
 (2/3) /home/misanjumn/tests/tests/avocado-misc-tests/perf/perf_script_bug.py:PerfScript.test_script_probe:  ERROR: No section: 'fedora' (0.01 s)
 (3/3) /home/misanjumn/tests/tests/avocado-misc-tests/perf/perf_test.py:Perftest.test_perf_test: STARTED
 (3/3) /home/misanjumn/tests/tests/avocado-misc-tests/perf/perf_test.py:Perftest.test_perf_test:  CANCEL: perf-debuginfo is needed for the test to be run (3.04 s)
RESULTS    : PASS 0 | ERROR 1 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 2
```

## With the patch
FAIL: 1
PASS: 2
```
 (1/3) /home/misanjumn/tests/tests/avocado-misc-tests/perf/perf_duplicate_probe.py:PerfDuplicateProbe.test_duplicate_probe: STARTED
 (1/3) /home/misanjumn/tests/tests/avocado-misc-tests/perf/perf_duplicate_probe.py:PerfDuplicateProbe.test_duplicate_probe:  PASS (0.75 s)
 (2/3) /home/misanjumn/tests/tests/avocado-misc-tests/perf/perf_script_bug.py:PerfScript.test_script_probe: STARTED
 (2/3) /home/misanjumn/tests/tests/avocado-misc-tests/perf/perf_script_bug.py:PerfScript.test_script_probe:  PASS (1.81 s)
 (3/3) /home/misanjumn/tests/tests/avocado-misc-tests/perf/perf_test.py:Perftest.test_perf_test: STARTED
 (3/3) /home/misanjumn/tests/tests/avocado-misc-tests/perf/perf_test.py:Perftest.test_perf_test:  FAIL: 4 Test failed (27.17 s)
RESULTS    : PASS 2 | ERROR 0 | FAIL 1 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
```

### perf_test.py
OS: `Fedora40`
command: `perf test`
Test: `perf/perf_test.py:Perftest.test_perf_test`
Without the patch: `CANCEL`
With the patch: `4 FAIL Test-cases`